backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.6
httpx>=0.24.0
python-dotenv>=1.0
psycopg2-binary>=2.9
dj-database-url>=2.3
gunicorn>=23.0
whitenoise>=6.9
